# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for tasks with JWT auth and spaces (personal + corporate). Register or login at /api/auth/*; use GET /api/auth/check to validate session. Send token as Authorization: Bearer \u0026lt;token\u0026gt; for /api/todos, /api/create, /api/spaces. Todos support space_id and assignee_id for corporate workspaces.",
        "title": "Task.grid API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List notifications for the current user. Sorted by created_at DESC.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "List notifications",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Max items (default 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by notification type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Include archived notifications (default false)",
                        "name": "include_archived",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.NotificationItem"
                            }
                        }
                    }
                }
            }
        },
        "/api/notifications/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Set read/unread and archive/unarchive for the notification. User must own the notification.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "notifications"
                ],
                "summary": "Update notification state",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "read, archived",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/notifications/{id}/archive": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Archive notification. User must own the notification.",
                "tags": [
                    "notifications"
                ],
                "summary": "Archive notification",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/api/spaces": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List spaces for the current user (personal + corporate where member). Ensures personal space exists.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spaces"
                ],
                "summary": "List spaces",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.SpaceListItem"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new corporate space. Caller becomes owner and admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spaces"
                ],
                "summary": "Create corporate space",
                "parameters": [
                    {
                        "description": "name",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateSpaceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SpaceListItem"
                        }
                    }
                }
            }
        },
...
        "/api/spaces/{id}/activity": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns activity log for corporate space. Only space admins can access.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spaces"
                ],
                "summary": "Get space activity",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Space ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Max items (default 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by event type",
                        "name": "event_type",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/handlers.SpaceActivityItem"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.CreateSpaceRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
...
        "handlers.NotificationItem": {
            "type": "object",
            "properties": {
                "archived_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "invitation_id": {
                    "type": "integer"
                },
                "read_at": {
                    "type": "string"
                },
                "space_id": {
                    "type": "integer"
                },
                "space_name": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "todo_id": {
                    "type": "integer"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "handlers.SpaceActivityItem": {
            "type": "object",
            "properties": {
                "actor_id": {
                    "type": "integer"
                },
                "actor_name": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "event_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "payload": {
                    "type": "object",
                    "additionalProperties": true
                },
                "space_id": {
                    "type": "integer"
                },
                "subject_name": {
                    "type": "string"
                },
                "subject_user_id": {
                    "type": "integer"
                },
                "todo_id": {
                    "type": "integer"
                }
            }
        },
...
        "handlers.SpaceListItem": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}